Move CLI subcommands from cmd/ai to pkg/cli
Move ls, send, kill, watch, run subcommands and parseSystemPrompt
from cmd/ai/ to pkg/cli/. cmd/ai/main.go is now a thin dispatch shell
that calls cli.* functions.

cmd/ai non-test lines: 5707 → 3276 (-43%)
Only the RPC application layer (rpcApp + handlers) remains in cmd/ai.

Tests moved: ls_test, kill_test, run_test, watch_test, watch_freeze_test,
helpers_test (was main_test). BuildRPCFlags tests extracted from
model_override_test to pkg/cli/run_flags_test.go.
