+++ pkg/cli/helpers_test.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"os"
@@ -7,26 +7,26 @@ import (
 
 func TestParseSystemPrompt(t *testing.T) {
 	// Test 1: Empty string
-	result := parseSystemPrompt("")
+	result := ParseSystemPrompt("")
 	if result != "" {
 		t.Errorf("Expected empty string, got %q", result)
 	}
 
 	// Test 2: Plain text (no @ prefix)
 	plainText := "You are a helpful assistant"
-	result = parseSystemPrompt(plainText)
+	result = ParseSystemPrompt(plainText)
 	if result != plainText {
 		t.Errorf("Expected %q, got %q", plainText, result)
 	}
 
 	// Test 3: @ prefix with empty path
-	result = parseSystemPrompt("@")
+	result = ParseSystemPrompt("@")
 	if result != "" {
 		t.Errorf("Expected empty string for @ with no path, got %q", result)
 	}
 
 	// Test 4: @ prefix with whitespace only
-	result = parseSystemPrompt("@   ")
+	result = ParseSystemPrompt("@   ")
 	if result != "" {
 		t.Errorf("Expected empty string for @ with whitespace, got %q", result)
 	}
@@ -37,13 +37,13 @@ func TestParseSystemPrompt(t *testing.T) {
 	if err := os.WriteFile(tmpFile, []byte(content), 0644); err != nil {
 		t.Fatalf("Failed to create temp file: %v", err)
 	}
-	result = parseSystemPrompt("@" + tmpFile)
+	result = ParseSystemPrompt("@" + tmpFile)
 	if result != content {
 		t.Errorf("Expected %q, got %q", content, result)
 	}
 
 	// Test 6: @ prefix with non-existent file
-	result = parseSystemPrompt("@/nonexistent/path/to/file.md")
+	result = ParseSystemPrompt("@/nonexistent/path/to/file.md")
 	if result != "" {
 		t.Errorf("Expected empty string for non-existent file, got %q", result)
 	}
@@ -57,7 +57,7 @@ func TestParseSystemPrompt(t *testing.T) {
 	if err := os.WriteFile(largeFile, largeContent, 0644); err != nil {
 		t.Fatalf("Failed to create large temp file: %v", err)
 	}
-	result = parseSystemPrompt("@" + largeFile)
+	result = ParseSystemPrompt("@" + largeFile)
 	if len(result) != 64*1024 {
 		t.Errorf("Expected 64KB, got %d bytes", len(result))
 	}
@@ -67,7 +67,7 @@ func TestParseSystemPrompt(t *testing.T) {
 	if err := os.WriteFile(tmpFile2, []byte("Content 2"), 0644); err != nil {
 		t.Fatalf("Failed to create temp file: %v", err)
 	}
-	result = parseSystemPrompt("@   " + tmpFile2)
+	result = ParseSystemPrompt("@   " + tmpFile2)
 	if result != "Content 2" {
 		t.Errorf("Expected 'Content 2', got %q", result)
 	}
