+++ pkg/cli/watch_freeze_test.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"strings"
